# backend/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "TTGO Backend API",
    "description": "A Express.js backend API for TTGO task management.",
    "version": "1.0.0",
    "contact": {
      "name": "Zachary Taylor",
      "email": "TG-Fullstack@outlook.com"
    } 
  },
  "servers": [
    {
      "url": "https://virtserver.swaggerhub.com/taylord/ttgoApi/1.0.0",
      "description": "SwaggerHub API Auto Mocking"
    },
    {
      "url": "https://localhost:3001/api/v1",
      "description": "Local backend API server"
    }
  ],
  "paths": {
...
    "/users/id": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "Return User by Id",
        "operationId": "getUsers",
        "responses": {
          "200": {
            "description": "Get User by Id",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "format": "uri",
                    "example": "http://10.0.0.225:8080"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "User"
        ],
        "operationId": "updateUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceRegistrationInfo"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully updated a User"
          }
        }
      },
      "put": {
        "tags": [
          "User"
        ],
        "operationId": "updateUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceRegistrationInfo"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully updated a User"
          }
        }
      }
    },
    "/projects": {
      "get": {
        "tags": [
          "Project"
        ],
        "description": "returns all projects",
        "operationId": "getProjects",
        "responses": {
          "200": {
            "description": "Successful fetch of projects",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "format": "uri",
                    "example": "http://10.0.0.225:8080"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Project"
        ],
        "operationId": "createProject",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceRegistrationInfo"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully created a Project."
          }
        }
      }
    },
    "/projects/${slug}": {
      "get": {
        "tags": [
          "Project"
        ],
        "description": "Return project",
        "operationId": "getProjects",
        "responses": {
          "200": {
            "description": "Successful fetch of projects",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "format": "uri",
                    "example": "http://10.0.0.225:8080"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Project"
        ],
        "operationId": "createProject",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceRegistrationInfo"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully created a Project."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeviceRegistrationInfo": {
        "type": "object",
        "properties": {
          "uri": {
            "type": "string",
            "format": "uri",
            "example": "http://10.0.0.220:8080"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "0729a580-2240-11e6-9eb5-0002a5d5c51b"
          }
        }
      }
    }
  }
}